add edge case for python built from source code when on windows

+++ crates/rattler_installs_packages/src/python_env/venv.rs
@@ -290,25 +290,30 @@ prompt = {}"#,
                 .parent()
                 .expect("cannot get system python parent folder");
             for bin_name in python_bins.into_iter() {
-                let original_python_bin = original_python_bin_dir.join("Lib/venv/scripts/nt").join(bin_name);
+                let original_python_bin = original_python_bin_dir.join(bin_name);
+                let original_python_scripts = original_python_bin_dir.join("Lib/venv/scripts/nt").join(bin_name);
+                let venv_python_bin = venv_bin.join(bin_name);
 
-                if original_python_bin.exists() {
-
-
-                    // edge case for when used python is built from source code
-                    // if basename.endswith('_d'):
-                    //     ext = '_d' + ext
-                    //     basename = basename[:-2]
-                    // if basename == 'python':
-                    //     basename = 'venvlauncher'
-                    // elif basename == 'pythonw':
-                    //     basename = 'venvwlauncher'
-
-
-                    let venv_python_bin = venv_bin.join(bin_name);
-                    if !venv_python_bin.exists() {
+                if original_python_bin.exists() && !venv_python_bin.exists(){
+                    if original_python_scripts.exists() {
                         copy_file(original_python_bin, &venv_python_bin)?;
+                    } else {
+                        // edge case for when used python is built from source code
+                        // or we use venv from build
+                        // let mut base_name = bin_name;
+                        let mut launcher_bin_name = bin_name.to_owned();
+                        
+                        if bin_name.contains("python") {
+                            launcher_bin_name = bin_name.replace("python", "venvlauncher");
+                        } else if bin_name.contains("pythonw") {
+                            launcher_bin_name = "venvwlauncher".to_owned();
+                        }
+                        let original_launcher_bin = original_python_bin_dir.join(launcher_bin_name);
+                        if original_launcher_bin.exists() {
+                            copy_file(original_python_bin, &venv_python_bin)?;
+                        }
                     }
+                    
                 }
             }
         }
